Ausleitung der Kräfteübersicht ergänzt. Anzahl AGTs wird erfasst.

+++ scripts/kraefteuebersicht.py
@@ -4,6 +4,7 @@ from tkinter import font
 import ttkbootstrap as ttk
 import customtkinter as ctk
 import datetime
+import fpdf
 from pymongo import ReturnDocument
 
 class Kraefteuebersicht(ttk.Frame):
@@ -16,11 +17,14 @@ class Kraefteuebersicht(ttk.Frame):
         self.label = ttk.Label(self, text='Kräfteübersicht', style='bolt')
         
         # Übersichtstabelle
-        self.headings = ['funkrufname', 'kraft', 'anmerkung']
+        self.headings = ['funkrufname', 'kraft', 'agt', 'anmerkung']
         self.table = ttk.Treeview(master=self, columns=self.headings, displaycolumns=self.headings, show='headings')
         for head in self.headings:
             self.table.heading(head, text=head.capitalize(), anchor='w')
-            self.table.column(head) # width=120, minwidth=100, stretch=False
+        self.table.column('funkrufname', width=140, minwidth=10, stretch=False)
+        self.table.column('kraft', width=80, minwidth=10, stretch=False)
+        self.table.column('agt', width=40, minwidth=10, stretch=False)
+        self.table.column('anmerkung', minwidth=10)
         self.table.bind('<<TreeviewSelect>>', self.item_selection)
         
         self.bearbeitungsmakse = Bearbeitungsmaske(self)
@@ -30,13 +34,16 @@ class Kraefteuebersicht(ttk.Frame):
         self.zf_ges = tk.IntVar(self, 0)
         self.gf_ges = tk.IntVar(self, 0)
         self.ms_ges = tk.IntVar(self, 0)
+        self.agt_ges = tk.IntVar(self, 0)
         self.fzg_ges = tk.IntVar(self, 0)
         self.anzeige_ges1 = tk.StringVar(self)
         self.anzeige_ges2 = tk.StringVar(self)
+        self.anzeige_ges3 = tk.StringVar(self, f'Atemschutzgeräteträger: {self.agt_ges}')
         
         self.label_frame = ttk.Frame(self)
         self.label_ges1 = ttk.Label(self.label_frame, textvariable=self.anzeige_ges1, font={'size':20})
         self.label_ges2 = ttk.Label(self.label_frame, textvariable=self.anzeige_ges2, font={'size':20})
+        self.label_ges3 = ttk.Label(self.label_frame, textvariable=self.anzeige_ges3, font={'size':16})
         self.f = font.Font(self.label_ges2, self.label_ges2.cget("font"))
         self.f.configure(underline=True)
         self.label_ges2.configure(font=self.f)
@@ -50,7 +57,9 @@ class Kraefteuebersicht(ttk.Frame):
         
         self.label_ges1.grid(row=1, column=1, sticky='ne', pady=10, padx=(5,0))
         self.label_ges2.grid(row=1, column=2, sticky='nw', pady=10, padx=(0,5))
+        self.label_ges3.grid(row=1, column=3, sticky='ne', pady=10, padx=(30,5))
         
+        ctk.CTkButton(self.label_frame, text='Kräftübersicht ausleiten', command=self.create_report).grid(row=2, column=1, columnspan=3, sticky='n', pady=10, padx=(5,5))
         
         self.loop()      
         
@@ -61,7 +70,10 @@ class Kraefteuebersicht(ttk.Frame):
     
     def pack_me(self):
         self.pack(pady=5, padx=5, fill='both', expand=True)
-        
+    
+    def create_report(self):        
+        report = Bericht(self.db.krafte.find(), self.einstellungen.orga_name.get())
+            
     def fill_table(self):
         self.clear_table()
         
@@ -71,34 +83,40 @@ class Kraefteuebersicht(ttk.Frame):
         zf_ges = 0
         gf_ges = 0
         ms_ges = 0
+        agt_ges = 0
         fzg_ges = 0
         
         for i, kraft in enumerate(krafte):
             row_tag = 'even' if (i%2==0) else 'odd'
             values = list(kraft.values())
+            #print(values)
             funk = values[1]
             vf = values[2]
             zf = values[3]
             gf = values[4]
             ms = values[5]
-            anmerkung = values[6]
-            zeile = (funk, f'{vf}/{zf}/{gf}/{ms}', anmerkung)
+            agt = values[6]
+            anmerkung = values[7]
+            zeile = (funk, f'{vf}/{zf}/{gf}/{ms}', agt, anmerkung)
             self.table.insert(parent='', index='end', values=zeile, tags=(row_tag,))
             
             vf_ges += int(vf)
             zf_ges += int(zf)
             gf_ges += int(gf)
             ms_ges += int(ms)
+            agt_ges += int(agt)
             fzg_ges += 1
         
         self.vf_ges.set(vf_ges)
         self.zf_ges.set(zf_ges)
         self.gf_ges.set(gf_ges)
         self.ms_ges.set(ms_ges)
+        self.agt_ges.set(agt_ges)
         self.fzg_ges.set(fzg_ges)
         gesamt = vf_ges+zf_ges+gf_ges+ms_ges
         self.anzeige_ges1.set(f'{vf_ges} / {zf_ges} / {gf_ges} / {ms_ges} / ')
         self.anzeige_ges2.set(gesamt)
+        self.anzeige_ges3.set(f'Atemschutzgeräteträger: {agt_ges}')
     
     def clear_table(self):
         for element in self.table.get_children():
@@ -111,7 +129,7 @@ class Kraefteuebersicht(ttk.Frame):
             kraft = [int(x) for x in values[1].split('/')]
             
             self.bearbeitungsmakse.anmerkung_entry.delete(0, 'end')
-            self.bearbeitungsmakse.anmerkung_entry.insert(0, values[2])
+            self.bearbeitungsmakse.anmerkung_entry.insert(0, values[3])
             
             self.bearbeitungsmakse.funkrufname_entry.delete(0, 'end')
             self.bearbeitungsmakse.funkrufname_entry.insert(0, values[0])
@@ -120,12 +138,8 @@ class Kraefteuebersicht(ttk.Frame):
             self.bearbeitungsmakse.zugfuhrer.current(kraft[1])
             self.bearbeitungsmakse.gruppenfuhrer.current(kraft[2])
             self.bearbeitungsmakse.mannschaft.current(kraft[3])
-            
-            #self.bearbeitungsmakse.btn_new.configure(state='disabel', fg_color='grey')
-        else:
-            pass
-            #self.bearbeitungsmakse.btn_new.configure(state='normal', fg_color=self.bearbeitungsmakse.btn_save.cget('fg_color'))
-
+            self.bearbeitungsmakse.agt.current(int(values[2]))
+        
 
 class Bearbeitungsmaske(ttk.Frame):
     def __init__(self, parent):
@@ -134,29 +148,25 @@ class Bearbeitungsmaske(ttk.Frame):
         self.parent = parent
 
         # Eingabe Elemente
-        #self.funkrufname = tk.StringVar(self)
         self.funkrufname_entry = ctk.CTkEntry(self, placeholder_text='Funkrufname')
         
-        #self.anzahl_verbandsfuhrer = tk.IntVar(self, 0)
         self.verbandsfuhrer = ttk.Combobox(self, values=list(range(100)), width=8)
         self.verbandsfuhrer.current(0)
         
-        #self.anzahl_zugfuhrer = tk.IntVar(self, 0)
         self.zugfuhrer = ttk.Combobox(self, values=list(range(100)), width=8)
         self.zugfuhrer.current(0)
         
-        #self.anzahl_gruppenfuhrer = tk.IntVar(self, 0)
         self.gruppenfuhrer = ttk.Combobox(self, values=list(range(100)), width=8)
         self.gruppenfuhrer.current(0)
         
-        #self.anzahl_mannschaft = tk.IntVar(self, 0)
         self.mannschaft = ttk.Combobox(self, values=list(range(1000)), width=8)
         self.mannschaft.current(0)
         
-        #self.anmerkung = tk.StringVar(self)
+        self.agt = ttk.Combobox(self, values=list(range(1000)), width=8)
+        self.agt.current(0)
+        
         self.anmerkung_entry = ctk.CTkEntry(self, placeholder_text='Anmerkung')
         
-        #self.btn_new = ctk.CTkButton(self, text='Neu')
         self.btn_save = ctk.CTkButton(self, text='Speichern', command=self.speicher_eintrag)
         self.btn_delete = ctk.CTkButton(self, text='Löschen', command=self.entferne_eintrag)
         
@@ -171,11 +181,13 @@ class Bearbeitungsmaske(ttk.Frame):
         self.gruppenfuhrer.grid(row=4, column=5)
         ttk.Label(self, text='/').grid(row=4, column=6)
         self.mannschaft.grid(row=4, column=7)
-        ttk.Label(self, text='Anmerkung').grid(row=5, column=1, columnspan=7, sticky='w')
-        self.anmerkung_entry.grid(row=6, column=1, columnspan=7, sticky='we')
-        # self.btn_new.grid(row=7, column=1, pady=20)
-        self.btn_save.grid(row=7, column=3, pady=20)
-        self.btn_delete.grid(row=7, column=5, pady=20)
+        ttk.Label(self, text='Anzahl Atemschutzgeräteträger').grid(row=5, column=1, columnspan=7, sticky='w')
+        self.agt.grid(row=6, column=1)
+        ttk.Label(self, text='Anmerkung').grid(row=7, column=1, columnspan=7, sticky='w')
+        self.anmerkung_entry.grid(row=8, column=1, columnspan=7, sticky='we')
+        
+        self.btn_save.grid(row=9, column=3, pady=20)
+        self.btn_delete.grid(row=9, column=5, pady=20)
         
     
     def speicher_eintrag(self):
@@ -197,6 +209,9 @@ class Bearbeitungsmaske(ttk.Frame):
         ms = self.mannschaft.get()
         self.mannschaft.current(0)
         
+        agt = self.agt.get()
+        self.agt.current(0)
+        
         cnt = self.db.krafte.count_documents({"funkrufname": funk})
         if cnt>0:
             self.db.krafte.find_one_and_update(
@@ -207,6 +222,7 @@ class Bearbeitungsmaske(ttk.Frame):
                         'zf': zf,
                         'gf': gf,
                         'ms': ms,
+                        'agt': agt,
                         'anmerkung': anmerkung,
                         'datum': datetime.datetime.now()
                     }}, 
@@ -219,6 +235,7 @@ class Bearbeitungsmaske(ttk.Frame):
                 'zf': zf,
                 'gf': gf,
                 'ms': ms,
+                'agt': agt,
                 'anmerkung': anmerkung,
                 'datum': datetime.datetime.now()
             })
@@ -228,7 +245,77 @@ class Bearbeitungsmaske(ttk.Frame):
     def entferne_eintrag(self):
         funk = self.funkrufname_entry.get()
         
-        #ToDo: Abfrage zum löschen einfügen.
+        res = tk.messagebox.askquestion('Löschne', 'Wirklich löschen?')
         
-        self.db.krafte.delete_one({'funkrufname': funk})
-        self.parent.fill_table()
+        if res == 'yes':
+            self.db.krafte.delete_one({'funkrufname': funk})
+            self.parent.fill_table()
+
+
+class Bericht(fpdf.FPDF):
+    def __init__(self, eintrage, organisation = 'Feuerwehr Musterstadt', orientation = "portrait", unit = "mm", format = "A4", font_cache_dir = "DEPRECATED") -> None:
+        super().__init__(orientation, unit, format, font_cache_dir)
+        
+        self.jetzt = datetime.datetime.now()
+        self.jetzt_einfach = self.jetzt.strftime('%d.%m.%Y %H:%M')
+        self.jetzt_einsatz = self.jetzt.strftime('%d%H%M%b%y')
+        
+        self.ordner_name = 'kraefteuebersicht'
+        if not os.path.exists(self.ordner_name):
+            os.mkdir(self.ordner_name)
+        self.path = os.path.join(self.ordner_name, f'{self.jetzt_einsatz}.pdf')
+        
+        self.organisation = organisation
+        
+        self.add_page()
+        
+        # Tabelle
+        self.spalten = ['Funkrufname', 'Anmerkung', 'VF', 'ZF', 'GF', 'MS', 'Ges', 'AGT']
+        self.spalten_breite = [30, 100, 10, 10, 10, 10, 10, 10]
+        with self.table(col_widths=tuple(self.spalten_breite)) as table:
+            self.row = table.row()
+            for cell in self.spalten:
+                self.row.cell(cell)
+            for eintrag in eintrage:
+                row = table.row()
+                
+                vf, zf, gf, ms = int(eintrag['vf']), int(eintrag['zf']), int(eintrag['gf']), int(eintrag['ms'])
+                agt = int(eintrag['agt'])
+                ges = vf+zf+gf+ms
+                for cell in [eintrag['funkrufname'], eintrag['anmerkung'], eintrag['vf'], eintrag['zf'], eintrag['gf'], eintrag['ms'], str(ges), str(agt)]:
+                    row.cell(cell, align=fpdf.Align.L)
+        
+        #ToDo: Gesamtzahlen drucken
+        
+        # Save pdf
+        self.output(self.path)
+        
+    def header(self):
+        # Name der Organisation einfügen
+        self.set_y(14)
+        self.set_font(family="helvetica", style='B', size=16)
+        self.cell(text=self.organisation)
+        
+        # Einsatzdaten
+        self.set_font(style='', size=14)
+        self.set_y(10)
+        self.cell(0, 10, text='Kräfteübersicht', align=fpdf.Align.C)
+        
+        # Erstelldatum oben rechts einfügen
+        self.set_font(style="", size=10)
+        self.set_y(10)
+        self.cell(0, 10, text=self.jetzt_einfach, align=fpdf.Align.R)
+        self.set_y(15)
+        self.cell(0, 10, text=self.jetzt_einsatz, align=fpdf.Align.R)
+        self.set_y(35)
+
+    
+    def footer(self):
+        # Position cursor at 1.5 cm from bottom:
+        self.set_y(-15)
+        # Setting font: helvetica italic 8
+        self.set_font("helvetica", "", 8)
+        # Printing page number:
+        self.cell(0, 10, f"Kräfteübersicht {self.jetzt_einsatz}", align=fpdf.Align.L)
+        self.set_y(-15)
+        self.cell(0, 10, f"Seite {self.page_no()}/{{nb}}", align=fpdf.Align.R)
